Vivado: Add even unknown files
Add even unknown file types to a Vivado project, using the generic
`add_files` Vivado command. It is then up to Vivado to figure out how to
handle a file (or if it wants to handle the file at all). This makes
using edalize easier to use with file types which are yet unknown to
edalize: instead of patching edalize and waiting for all users to pick
up a new version, users can rely on Vivado to do the right thing, which
should work most of the time.