Behavior change in lstrip_blocks in Jinja 2.11.0
Jinja 2.11.0 got released a couple days ago, and it changed the behavior
of lstrip_blocks. This breaks some templates.

It's not yet clear if we misunderstood the lstrip_blocks documentation
(and things were never expected to work that way), or if Jinja2
introduced a breaking change.

Upstream bug report https://github.com/pallets/jinja/issues/1138