Start deprecating the argument parser in Edalize
The argument parser is used in two places in Edalize and this starts
the work of removing both cases

configure: Edalize will no longer accept arguments passed as a list
through configure. Users will instead have to set values in the default
key of the parameter.

run: Edalize will still accept and parse a list of arguments, but the
preferred method is to send a dict of key/value pairs with runtime
values of parameters that will override the default values from the
configure stage