Add support for SDF files
These have the property sdf_type to identify if the file should
be copied to the working directory for use by $sdf_annotate and
friends or if the backend should add it as a min, max, or typical
corner. The property sdf_instance identifies a design instance
where the SDF file should be applied.