SymbiYosys support
The SymbiYosys tool is a little tricky to support because the .sby
file that it uses for configuration contains a list of files, as well
as a list of commands. The user is going to want to edit the latter,
but the former needs to come from fusesoc or similar.

This patch assumes that the .sby file will be a Jinja2 template.
There's detailed documentation about the format in the Symbiyosys
class's _description (which should appear with --help).

A minimal example would be something like this:

    [options]
    mode bmc
    depth 3

    [script]
    {{"-sv"|gen_reads}}
    prep -top my_top_module

    [files]
    {{files}}

When run, the 'configure' step interpolates the .sby configuration
script, writing to <work_root>/test.sby. It also dumps a list of RTL
file paths (to <work_root>/files.txt) and include directories (to
<work_root>/incdirs.txt). This is handy if you need to run some sort
of preprocessing step (sv2v?) on your input files to massage them into
a form that Yosys will support.

The 'build' step does nothing.

The 'run' step runs 'sby'. The backend has a single parameter,
"tasknames". This defaults to empty, in which case, the command is
something like

    sby -d build test.sby

However, if you want multiple tasks in your config file, you'll need
to specify which ones to run. "tasknames" let's you pass them, ending
up with a command like

    sby -d build test.sby task1 task2

Note: This adds a file to the mock file list in
edalize_common.py (because the symbiyosys backend needs a .sby file).
The Vivado test needed regenerating with GOLDEN_RUN=1; the others
seemingly didn't need any changes.