Take class docs from both __init__(), and the class docstring.
Sphinx autodoc, by default, expects the docstring for a class, which is
equal to the docstring of the constructor, to be in the class. That's
slightly odd at times, since the arguments of the constructor then need
to be documented in the class. Use a setting that enables splitting of
the class documentation: the class documentation can stay, and
constructor arguments can be documented in the constructor.