Add configuration options for tool output
Introduce a new verbose flag to enable all tool output to be written
to the terminal. This flag is on by default to correspond to the old
behaviour. Turning off the verbose flag will still produce output
during the run stage.

Also add stdout and stderr objects that can be passed to the edatool
class to redirect stdout/stderr to streams instead of the terminal