Add support for custom launchers
This add suspport for setting an EDALICE_LAUNCHER environment variable
to specify a custom command to be run with the rest of the command-line
passed as arguments.

The main purpose of this is to let the custom launcher add a prefix command
to the command line. There are several use cases for this such as running
tools from containers or starting a license server prior to running a tool.

This fix covers a subset of the backends. More backends will be supported
in time and there will eventually be an option for setting the launcher as
a command-line option rather than through environment variables.