Initial commit on Openlane support.
The absolute bare minimum to run Openlane on a design through Edalize.
Generates a config.tcl with the information needed by the tool and
a Makefile for launching it.
Current templates includes several assumptions suitable for building
SERV for later integration in another project - this is just for testing
purposes.