Add support for top level that includes a library specification
Fixes #229.

Many tools allow the top level unit in a non-default library to be
specified using the syntax my_lib.my_unit. GHDL doesn't and instead
requires the library to be specified separately with the ``--work`` and
``--workdir`` options. This converts the common dot notation to GHDLs
syntax.

This change also reduces the number of GHDL commands by combining all
source files for a library into a single command rather than using a
GHDL command for every file.