Avoid warning from deprecated NumPy float alias
Tests using Python 3.7+ report the following warning:
```
  /home/runner/work/edalize/edalize/.tox/py/lib/python3.7/site-packages/edalize/quartus_reporting.py:112:
DeprecationWarning: `np.float` is a deprecated alias for the builtin
`float`. To silence this warning, use `float` by itself. Doing this will
not modify any behavior and is safe. If you specifically wanted the
numpy scalar type, use `np.float64` here.
  Deprecated in NumPy 1.20; for more details and guidance:
https://numpy.org/devdocs/release/1.20.0-notes.html#deprecations
    freq.str.split(expand=True)[0].astype(np.float).to_dict()
```

There are no other explicit uses of NumPy in this or other files, so the
import and ``setup.py`` requirement can also be removed. NumPy will
still be pulled in as a Pandas dependency but Edalize doesn't explicitly
need it.