Fix Sphinx parse error in Vivado documentation
The CAPI2 documentation and the edalize documentation are built
automatically from contents in the backends, and descriptions must
therefore be valid reStructuredText. Fix a case where this wasn't true.

Sphinx reported (in a FuseSoC build) due to the `*` not being followed
by another `*`:

```
capi2.rst:661:Inline emphasis start-string without end-string.
```