Initial attempt to make Quartus use the command generator
This passes the tests, though the Jinja filtering previously used
was complex enough that problems may appear in use.

This attempts to duplicate the previous Makefiles, though that may not
be good style with the command generator. This required modifying the
header to include Makefile variables since there's not yet support for
variables in the class