Fix test_edam_hook
The test was intended to execute a script which returns a non-zero exit
code, and assert that the execution failed. The execution of the script
actually failed and the test passed, but for the wrong reason: the test
script was at the wrong place.

Put the file into the right place, and strengthen the exception check.