Refactor and introduce get_tool_options in flow
Before this change, there was no meaningful way to use flow options
because it was not possible to query or set tool options for tools
outside of the ones in the static FLOW list.

This patch separates the process of applying options into two steps.
A user can first use get_flow_options to get flow options and then
use get_tool_options with the preferred flow options set, to get
all tool options for a configured flow.

The redefinition of get_flow_options to only return flow options
(and not also tool options like previously) is an API break but will
hopefully not cause too much trouble.