Stop passing args through the run function
All runtime arguments are set in the EDAM description. args passed
through the run function are not used. We still need to keep an
optional args=None argument in the flows since that is an external
API. This will be removed after leaving some time for users (e.g.
FuseSoC) to stop passing args.