Support simulation tag in Vivado
By setting the simulation tag on an input HDL file, it will be marked
as simulation-only which means that any change to this file will not
invalidate the netlist or implementation. The files marked with simulation
tags are also forwarded in the tool output EDAM instead of being consumed
by Vivado.