verilator: Fix dependencies on testbench files in makefile
The build is a two-stage process. First, the RTL files are "verilated"
into C++ files and a <project>.mk file is created. Then, the verilated
C++ files together with the testbench are compiled into an executable.

The Makefile generated by Edalize had a dependency on the testbench
files for the first step, but the testbench files are really only used
in the second step, so there was nothing to pick up the changes.