Support passing `vlogparam`s to Genus
_Note_: this is written from the viewpoint of a FuseSoC user, but this
applies to all usages of edalize itself.

FuseSoC already accepts `vlogparam`s for the `genus` backend/tool, but
does not write them to any TCL-file or other way to access it from the
user-defined Genus script. Therefore one could not easily add parameters
to the top-level module when synthesizing with Genus, effectively ren-
dering them unusable.

In order to resolve this issue, this commit creates new variable called
`ELABORATE_PARAMETERS`, which is set unconditionally. Its value is
constructed in such a way, that the value can be passed directly to the
`elaborate`-command inside the user-supplied Genus TCL script like so:
```tcl
elaborate -parameters "$ELABORATE_PARAMETERS" "$TOP_MODULE"
```
This way, the user-supplied TCL-script can be kept generic (without the
need for special-casing parameter) and FuseSoC makes use of the defined
parameters.